Add `nmodes` parameter for setting/getting extents.
For getting and setting both extents and strides, have the user provide an additional `nmodes` parameter which is the maximum number of values to return in the provided extent/stride array. If the number of modes provided is smaller than the number of modes in the tensor, truncate the returned information and return an error.

Fixes #13.