Fix destructors and add some constructors.
Add constructors for TAPP_attr, TAPP_handle, and TAPP_executor. The latter make use of the former to provide context and configuration for construction. For all destructors, pass a pointer to the object handle so that it can be reset to NULL or another neutral value. This prevents double-free, use-after-free, etc.